UPDATES AND BUG FIXES
Completely reworked how Switcher does the hiding and showing of stages. It now implements a ChangeEvent on the stages focusedProperty to streamline that process and remove dependency on a thread.

Created new methods for assigning the hiddenIfLostFocus behavior, depreciated former methods for this task

Depreciated the setVisible method and created the hide and unHide methods.